chore: Update package.json and vercel.json
This commit updates the package.json file to include a new dependency for "@inertiajs/inertia" with version "^0.11.1". It also adds a new "build" section in the package.json file with an "env" object containing the "NODE_ENV" set to "production" and a "command" set to "npm run build". Additionally, the vercel.json file is modified to remove the "scripts" section and add a new "build" section.